+++ src/screens/game_screen/square.rs
@@ -3,13 +3,13 @@ use ggez::graphics::Color;
 
 #[derive(Debug, Clone)]
 pub struct Square {
-    pub row: u8,
-    pub column: u8,
+    pub row: u16,
+    pub column: u16,
     pub color: Color,
 }
 
 impl Square {
-    pub fn new(row: u8, column: u8, color: Color) -> Square {
+    pub fn new(row: u16, column: u16, color: Color) -> Square {
         Square { row, column, color }
     }
 }
